+++ src/main/java/com/xcale/challenge/dto/GroupResponseDTO.java
@@ -1,7 +1,11 @@
 package com.xcale.challenge.dto;
 
+import lombok.Builder;
+import lombok.Data;
 import org.springframework.http.HttpStatus;
 
+@Data
+@Builder
 public class GroupResponseDTO {
 
     HttpStatus status;